\name{mpr_templates-package}
\alias{mpr_templates-package}
\alias{mpr_templates}
\docType{package}
\title{
\packageTitle{mpr_templates}
}
\description{
\packageDescription{mpr_templates}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mpr_templates}
\packageIndices{mpr_templates}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{mpr_templates}

Maintainer: \packageMaintainer{mpr_templates}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
